Add MultipleEqSlotUI prototype

+++ Giera/AbstractEqSlot.h
@@ -16,10 +16,10 @@ protected:
 	array <bool, (int)ItemTypes::COUNT> isAcceptedItemType;
 public:
 	AbstractEqSlot(vector <ItemTypes>& acceptedItemTypes);
-	virtual bool isAccepted(double x, double y, shared_ptr<AbstractItem> item) = 0;
-	virtual void insertAcceptedItem(double x, double y, shared_ptr<AbstractItem> item) = 0;
-	virtual optional<shared_ptr<AbstractItem>> removeItem(double x, double y) = 0;
-	virtual optional<shared_ptr<AbstractItem>> getItem(double x, double y) = 0;
+	virtual bool isAccepted(int x, int y, shared_ptr<AbstractItem> item) = 0;
+	virtual void insertAcceptedItem(int x, int y, shared_ptr<AbstractItem> item) = 0;
+	virtual optional<shared_ptr<AbstractItem>> removeItem(int x, int y) = 0;
+	virtual optional<shared_ptr<AbstractItem>> getItem(int x, int y) = 0;
 	virtual unique_ptr <EqSlotUIElement> generateUIElement(Rect <fr_pos_t> relRect,
 		UIElement* parent, shared_ptr <InventoryInputHandler> inventoryInputHandler) = 0;
 	virtual void updateUIElementItems(EqSlotUIElement* eqSlotUIElement,
